requirements.txt
fastapi~=0.115.13
uvicorn~=0.34.3
pydantic~=2.11.7
pydantic-settings~=2.10.0
SQLAlchemy~=2.0.41
psycopg-binary~=3.2.9
PyYAML~=6.0.2
orjson~=3.10.18
openai~=1.90.0
google-genai~=1.21.1
ollama~=0.5.1
tenacity~=8.5.0
httpx~=0.28.1
requests~=2.32.4
jsonschema~=4.24.0
google-api-core~=2.25.1
